+++ common/predictive-text/unit_tests/in_browser/browser-test.sh
@@ -32,7 +32,7 @@ get_browser_set_for_OS ( ) {
     if [ $os_id = "mac" ]; then
         BROWSERS="--browsers Firefox,Chrome,Safari"
     elif [ $os_id = "win" ]; then
-        BROWSERS="--browsers Chrome,IE"
+        BROWSERS="--browsers Chrome"
     else
         BROWSERS="--browsers Firefox,Chrome"
     fi