+++ web/source/text/inputProcessor.ts
@@ -1,4 +1,9 @@
+// Defines a 'polyfill' of sorts for NPM's events module
+/// <reference path="../includes/events.ts" />
+/// <reference path="../includes/lmlayer.ts" />
+/// <reference path="../includes/lmMsgs.d.ts" />
 /// <reference path="keyboardProcessor.ts" />
+/// <reference path="prediction/languageProcessor.ts" />
 
 namespace com.keyman.text {
   export class InputProcessor {
@@ -6,7 +11,8 @@ namespace com.keyman.text {
       baseLayout: 'us'
     }
 
-    kbdProcessor: KeyboardProcessor;
+    private kbdProcessor: KeyboardProcessor;
+    private lngProcessor: prediction.LanguageProcessor;
 
     constructor(options?: ProcessorInitOptions) {
       if(!options) {
@@ -14,6 +20,11 @@ namespace com.keyman.text {
       }
 
       this.kbdProcessor = new KeyboardProcessor(options);
+      this.lngProcessor = new prediction.LanguageProcessor();
+    }
+
+    public get languageProcessor(): prediction.LanguageProcessor {
+      return this.lngProcessor;
     }
 
     public get keyboardProcessor(): KeyboardProcessor {
@@ -36,6 +47,10 @@ namespace com.keyman.text {
       this.resetContext();
     }
 
+    public get activeModel(): prediction.ModelSpec {
+      return this.languageProcessor.activeModel;
+    }
+
         /**
      * Simulate a keystroke according to the touched keyboard button element
      *
@@ -43,13 +58,12 @@ namespace com.keyman.text {
      * 
      * @param       {Object}      e      The abstracted KeyEvent to use for keystroke processing
      */
-    processKeyEvent(keyEvent: KeyEvent): boolean {
-      let keyman = com.keyman.singleton;
+    processKeyEvent(keyEvent: KeyEvent) {
       let formFactor = keyEvent.device.formFactor;
 
       // Determine the current target for text output and create a "mock" backup
       // of its current, pre-input state.  Current, long-existing assumption - it's DOM-backed.
-      let outputTarget = keyEvent.Ltarg as dom.targets.OutputTarget;
+      let outputTarget = keyEvent.Ltarg;
       let fromOSK = keyEvent.isSynthetic;
 
       // The default OSK layout for desktop devices does not include nextlayer info, relying on modifier detection here.
@@ -71,14 +85,14 @@ namespace com.keyman.text {
       // If suggestions exist AND space is pressed, accept the suggestion and do not process the keystroke.
       // If a suggestion was just accepted AND backspace is pressed, revert the change and do not process the backspace.
       // We check the first condition here, while the prediction UI handles the second through the try__() methods below.
-      if(keyman.modelManager.enabled) {
+      if(this.languageProcessor.isActive) {
         // The following code relies on JS's logical operator "short-circuit" properties to prevent unwanted triggering of the second condition.
 
         // Can the suggestion UI revert a recent suggestion?  If so, do that and swallow the backspace.
-        if((keyEvent.kName == "K_BKSP" || keyEvent.Lcode == Codes.keyCodes["K_BKSP"]) && keyman.modelManager.tryRevertSuggestion()) {
+        if((keyEvent.kName == "K_BKSP" || keyEvent.Lcode == Codes.keyCodes["K_BKSP"]) && this.languageProcessor.tryRevertSuggestion()) {
           return;
           // Can the suggestion UI accept an existing suggestion?  If so, do that and swallow the space character.
-        } else if((keyEvent.kName == "K_SPACE" || keyEvent.Lcode == Codes.keyCodes["K_SPACE"]) && keyman.modelManager.tryAcceptSuggestion('space')) {
+        } else if((keyEvent.kName == "K_SPACE" || keyEvent.Lcode == Codes.keyCodes["K_SPACE"]) && this.languageProcessor.tryAcceptSuggestion('space')) {
           return;
         }
       }
@@ -99,7 +113,7 @@ namespace com.keyman.text {
 
         // If we're performing a 'default command', it's not a standard 'typing' event - don't do fat-finger stuff.
         // Also, don't do fat-finger stuff if predictive text isn't enabled.
-        if(keyman.modelManager.enabled && !ruleBehavior.triggersDefaultCommand) {
+        if(this.languageProcessor.isActive && !ruleBehavior.triggersDefaultCommand) {
           // Note - we don't yet do fat-fingering with longpress keys.
           if(keyEvent.keyDistribution && keyEvent.kbdLayer) {
             let activeLayout = this.activeKeyboard.layout(keyEvent.device.formFactor);
@@ -114,8 +128,7 @@ namespace com.keyman.text {
                 continue;
               }
 
-              let altEvent = osk.PreProcessor._GetClickEventProperties(altKey, outputTarget.getElement());
-              altEvent.Ltarg = mock;
+              let altEvent = altKey.constructKeyEvent(this.keyboardProcessor, mock, keyEvent.device);
               let alternateBehavior = this.keyboardProcessor.processKeystroke(altEvent, mock);
               if(alternateBehavior) {
                 // TODO: if alternateBehavior.beep == true, set 'p' to 0.  It's a disallowed key sequence,
@@ -137,30 +150,12 @@ namespace com.keyman.text {
         // by the actual keystroke occur.
         ruleBehavior.finalize(this.keyboardProcessor);
 
-        // If the transform isn't empty, we've changed text - which should produce a 'changed' event in the DOM.
-        //
-        // TODO:  This check should be done IN a dom module, not here in web-core space.  This place is closer 
-        //        to that goal than it previously was, at least.
-        let ruleTransform = ruleBehavior.transcription.transform;
-        if(ruleTransform.insert != "" || ruleTransform.deleteLeft > 0 || ruleTransform.deleteRight > 0) {
-          if(outputTarget.getElement() == dom.DOMEventHandlers.states.activeElement) {
-            dom.DOMEventHandlers.states.changed = true;
-          }
-        }
-
         // -- All keystroke (and 'alternate') processing is now complete.  Time to finalize everything! --
         
         // Notify the ModelManager of new input - it's predictive text time!
         ruleBehavior.transcription.alternates = alternates;
         // Yes, even for ruleBehavior.triggersDefaultCommand.  Those tend to change the context.
-        keyman.modelManager.predict(ruleBehavior.transcription);
-
-        // KMEA and KMEI (embedded mode) use direct insertion of the character string
-        if(keyman.isEmbedded) {
-          // A special embedded callback used to setup direct callbacks to app-native code.
-          keyman['oninserttext'](ruleTransform.deleteLeft, ruleTransform.insert, ruleTransform.deleteRight);
-          keyman.refreshElementContent(outputTarget.getElement());
-        }
+        ruleBehavior.predictionPromise = this.languageProcessor.predict(ruleBehavior.transcription);
 
         // Text did not change (thus, no text "input") if we tabbed or merely moved the caret.
         if(!ruleBehavior.triggersDefaultCommand) {
@@ -178,16 +173,12 @@ namespace com.keyman.text {
       //        once THOSE are properly relocated.  (They're too DOM-heavy to remain in web-core.)
 
       // Only return true (for the eventual event handler's return value) if we didn't match a rule.
-      return ruleBehavior == null;
+      return ruleBehavior;
     }
 
     public resetContext() {
-      let keyman = com.keyman.singleton;
       this.keyboardProcessor.resetContext();
-
-      if(keyman.modelManager) {
-        keyman.modelManager.invalidateContext();
-      }
+      this.languageProcessor.invalidateContext();
     }
   }
 }